+++ .prettierrc
@@ -1,4 +1,8 @@
 {
-  "singleQuote": true,
-  "trailingComma": "all"
+  "semi": true,
+  "trailingComma": "all",
+  "singleQuote": false,
+  "printWidth": 180,
+  "tabWidth": 2,
+  "arrowParens": "avoid"
 }
