+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "nestjs-boilerplate",
+  "name": "api-coopartilhar",
   "version": "0.0.1",
   "description": "",
   "author": "",
@@ -32,77 +32,76 @@
     "prepare": "is-ci || husky install"
   },
   "dependencies": {
-    "@aws-sdk/client-s3": "3.360.0",
+    "@aws-sdk/client-s3": "^3.576.0",
     "@nestjs/common": "^10.3.8",
     "@nestjs/config": "^3.2.2",
     "@nestjs/core": "^10.3.8",
-    "@nestjs/devtools-integration": "0.1.5",
-    "@nestjs/jwt": "10.1.0",
-    "@nestjs/passport": "10.0.0",
-
-    "@nestjs/platform-express": "10.3.8",
+    "@nestjs/devtools-integration": "^0.1.6",
+    "@nestjs/jwt": "^10.2.0",
+    "@nestjs/passport": "^10.0.3",
+    "@nestjs/platform-express": "^10.3.8",
     "@nestjs/swagger": "^7.3.1",
     "@nestjs/typeorm": "^10.0.2",
-    "@types/multer-s3": "^3.0.0",
-    "apple-signin-auth": "1.7.5",
+    "@types/multer-s3": "^3.0.3",
+    "apple-signin-auth": "^1.7.6",
     "aws-xray-sdk": "^3.6.0",
-    "bcryptjs": "2.4.3",
-    "class-transformer": "0.5.1",
-    "class-validator": "0.14.0",
-    "fb": "2.0.0",
-    "google-auth-library": "8.9.0",
-    "handlebars": "4.7.7",
+    "bcryptjs": "^2.4.3",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.14.1",
+    "fb": "^2.0.0",
+    "google-auth-library": "^8.9.0",
+    "handlebars": "^4.7.8",
     "ms": "^2.1.3",
-    "multer": "1.4.4",
-    "multer-s3": "3.0.1",
-    "nestjs-i18n": "^10.2.6",
-    "nodemailer": "6.9.9",
-    "passport": "0.6.0",
-    "passport-anonymous": "1.0.1",
-    "passport-jwt": "4.0.1",
-    "pg": "8.11.1",
-    "reflect-metadata": "0.1.13",
-    "rimraf": "5.0.1",
-    "rxjs": "7.8.1",
-    "source-map-support": "0.5.21",
-    "swagger-ui-express": "5.0.0",
-    "twitter": "1.7.1",
-    "typeorm": "0.3.17"
+    "multer": "^1.4.4",
+    "multer-s3": "^3.0.1",
+    "nestjs-i18n": "^10.4.5",
+    "nodemailer": "^6.9.13",
+    "passport": "^0.6.0",
+    "passport-anonymous": "^1.0.1",
+    "passport-jwt": "^4.0.1",
+    "pg": "^8.11.5",
+    "reflect-metadata": "^0.1.14",
+    "rimraf": "^5.0.7",
+    "rxjs": "^7.8.1",
+    "source-map-support": "^0.5.21",
+    "swagger-ui-express": "^5.0.0",
+    "twitter": "^1.7.1",
+    "typeorm": "^0.3.20"
   },
   "devDependencies": {
-    "@nestjs/cli": "10.1.7",
-    "@nestjs/schematics": "10.0.1",
-    "@nestjs/testing": "10.0.5",
-    "@types/bcryptjs": "2.4.2",
-    "@types/express": "4.17.17",
-    "@types/facebook-js-sdk": "3.3.6",
-    "@types/jest": "29.5.2",
-    "@types/ms": "^0.7.31",
-    "@types/multer": "1.4.7",
-    "@types/node": "18.16.19",
-    "@types/passport-anonymous": "1.0.3",
-    "@types/passport-jwt": "3.0.9",
-    "@types/supertest": "2.0.12",
-    "@types/twitter": "1.7.1",
-    "@typescript-eslint/eslint-plugin": "5.61.0",
-    "@typescript-eslint/parser": "5.61.0",
-    "env-cmd": "10.1.0",
-    "eslint": "8.44.0",
-    "eslint-config-prettier": "8.8.0",
-    "eslint-plugin-import": "2.27.5",
-    "eslint-plugin-prettier": "4.2.1",
-    "husky": "8.0.3",
-    "hygen": "6.2.11",
-    "is-ci": "3.0.1",
-    "jest": "29.6.1",
-    "prettier": "2.8.8",
-    "supertest": "6.3.3",
-    "ts-jest": "29.1.1",
-    "ts-loader": "9.4.4",
-    "ts-node": "10.9.1",
-    "tsconfig-paths": "4.2.0",
-    "tslib": "2.6.0",
-    "typescript": "5.1.6"
+    "@nestjs/cli": "^10.3.2",
+    "@nestjs/schematics": "^10.1.1",
+    "@nestjs/testing": "^10.3.8",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/express": "^4.17.21",
+    "@types/facebook-js-sdk": "^3.3.11",
+    "@types/jest": "^29.5.12",
+    "@types/ms": "^0.7.34",
+    "@types/multer": "^1.4.11",
+    "@types/node": "^18.19.33",
+    "@types/passport-anonymous": "^1.0.5",
+    "@types/passport-jwt": "^3.0.13",
+    "@types/supertest": "^2.0.16",
+    "@types/twitter": "^1.7.4",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
+    "env-cmd": "^10.1.0",
+    "eslint": "^9.2.0",
+    "eslint-config-prettier": "^8.10.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "husky": "^8.0.3",
+    "hygen": "^6.2.11",
+    "is-ci": "^3.0.1",
+    "jest": "^29.7.0",
+    "prettier": "^2.8.8",
+    "supertest": "^6.3.4",
+    "ts-jest": "^29.1.2",
+    "ts-loader": "^9.5.1",
+    "ts-node": "^10.9.2",
+    "tsconfig-paths": "^4.2.0",
+    "tslib": "^2.6.2",
+    "typescript": "^5.4.5"
   },
   "jest": {
     "moduleFileExtensions": [
@@ -122,6 +121,6 @@
     "testEnvironment": "node"
   },
   "engines": {
-    "node": ">=16.0.0"
+    "node": ">=20.0.0"
   }
 }
