+++ src/@types/index.d.ts
@@ -3,7 +3,7 @@
 declare global {
   namespace Express {
     declare interface Request {
-      hash: string;
+      hash?: string;
     }
   }
 }
