--- src/app.module.ts
@@ -1,113 +1,0 @@
-import { Module } from '@nestjs/common';
-import databaseConfig from './config/database.config';
-import authConfig from './config/auth.config';
-import appConfig from './config/app.config';
-import mailConfig from './config/mail.config';
-import fileConfig from './config/file.config';
-import facebookConfig from './config/facebook.config';
-import googleConfig from './config/google.config';
-import twitterConfig from './config/twitter.config';
-import appleConfig from './config/apple.config';
-import path from 'path';
-
-import { UsersModule } from './modules/user/users.module';
-import { FilesModule } from './modules/files/files.module';
-import { AuthModule } from './modules/auth/auth.module';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthAppleModule } from './modules/auth/social/apple/auth-apple.module';
-import { AuthFacebookModule } from './modules/auth/social/facebook/auth-facebook.module';
-import { AuthGoogleModule } from './modules/auth/social/google/auth-google.module';
-import { AuthTwitterModule } from './modules/auth/social/twitter/auth-twitter.module';
-import { I18nModule } from 'nestjs-i18n/dist/i18n.module';
-import { HeaderResolver } from 'nestjs-i18n';
-import { TypeOrmConfigService } from './database/typeorm-config.service';
-import { ForgotModule } from './modules/forgot/forgot.module';
-import { MailModule } from './mail/mail.module';
-import { HomeModule } from './modules/home/home.module';
-import { DataSource, DataSourceOptions } from 'typeorm';
-import { AllConfigType } from './config/config.type';
-import { SessionModule } from './modules/session/session.module';
-import { MailerModule } from './mailer/mailer.module';
-import { DevtoolsModule } from '@nestjs/devtools-integration';
-import { APP_INTERCEPTOR } from '@nestjs/core';
-import { LoggingInterceptor } from './infrastructure/interceptors/logging.interceptor';
-import { AppController } from './modules/app/app.controller';
-import { AppService } from './modules/app/app.service';
-
-@Module({
-  imports: [
-    DevtoolsModule.register({
-      http: process.env.NODE_ENV !== 'production',
-    }),
-    ConfigModule.forRoot({
-      isGlobal: true,
-      load: [
-        databaseConfig,
-        authConfig,
-        appConfig,
-        mailConfig,
-        fileConfig,
-        facebookConfig,
-        googleConfig,
-        twitterConfig,
-        appleConfig,
-      ],
-      envFilePath: ['.env'],
-    }),
-    TypeOrmModule.forRootAsync({
-      useClass: TypeOrmConfigService,
-      dataSourceFactory: async (options: DataSourceOptions) => {
-        return new DataSource(options).initialize();
-      },
-    }),
-    I18nModule.forRootAsync({
-      useFactory: (configService: ConfigService<AllConfigType>) => ({
-        fallbackLanguage: configService.getOrThrow('app.fallbackLanguage', {
-          infer: true,
-        }),
-        loaderOptions: { path: path.join(__dirname, '/i18n/'), watch: true },
-      }),
-      resolvers: [
-        {
-          use: HeaderResolver,
-          useFactory: (configService: ConfigService<AllConfigType>) => {
-            return [
-              configService.get('app.headerLanguage', {
-                infer: true,
-              }),
-            ];
-          },
-          inject: [ConfigService],
-        },
-      ],
-      imports: [ConfigModule],
-      inject: [ConfigService],
-    }),
-    UsersModule,
-    FilesModule,
-    AuthModule,
-    AuthFacebookModule,
-    AuthGoogleModule,
-    AuthTwitterModule,
-    AuthAppleModule,
-    ForgotModule,
-    SessionModule,
-    MailModule,
-    MailerModule,
-    HomeModule,
-    AppModule,
-  ],
-  controllers: [AppController],
-  providers: [
-    AppService,
-    { provide: APP_INTERCEPTOR, useClass: LoggingInterceptor },
-    /**
-     * In a future scenario where it is interesting to use request tracing via xRay in AWS, simply enable this logging
-     */
-    /**
-     * { provide: APP_INTERCEPTOR, useClass: XRayInterceptor }
-     */
-  ],
-})
-export class AppModule {}
